# Conflicts:
#	CodeVein_0.1/Client/Codes/GunGenji.cpp
#	CodeVein_0.1/Client/Codes/MonsterUI.cpp
#	CodeVein_0.1/Client/Codes/PoisonButterfly.cpp
#	CodeVein_0.1/Client/Codes/Scene_Stage_Training.cpp
#	CodeVein_0.1/Client/Codes/SwordGenji.cpp
#	CodeVein_0.1/Client/Default/Client.vcxproj.filters
#	CodeVein_0.1/MonsterTool/Bin/Tool_Monster.exe

+++ CodeVein_0.1/Client/Codes/Expendables_Inven.cpp
@@ -32,7 +32,9 @@ HRESULT CExpendables_Inven::Ready_GameObject(void * pArg)
 
 	m_fPosX = WINCX * 0.3f;
 	m_fPosY = WINCY * 0.5f;
-	m_fSizeX = 352.f;
+
+	m_fSizeX = 280.f;
+
 	m_fSizeY = 471.f;
 
 	m_fViewZ = 1.f;
@@ -47,21 +49,26 @@ HRESULT CExpendables_Inven::Ready_GameObject(void * pArg)
 		for (_uint j = 0; j < 5; ++j)
 		{
 			pDesc = new CUI::UI_DESC;
-			pDesc->fPosX = m_fPosX - 120.f + 60.f * j;
-			pDesc->fPosY = m_fPosY - 130.f + 60.f * i;
+
+			pDesc->fPosX = m_fPosX - 103.f + 52.f * j;
+			pDesc->fPosY = m_fPosY - 130.f + 52.f * i;
 			pDesc->fSizeX = 50.f;
 			pDesc->fSizeY = 50.f;
 			g_pManagement->Add_GameObject_ToLayer(L"GameObject_ExpendSlot", SCENE_STAGE, L"Layer_ExpendSlot", pDesc);
 			pExSlot = static_cast<CExpendables_Slot*>(g_pManagement->Get_GameObjectBack(L"Layer_ExpendSlot", SCENE_STAGE));
 
 			m_vecSlot.push_back(pExSlot);
+			m_vecUI_DESC.push_back(pDesc);
 		}
 	}
 
-	Add_Expendables(CExpendables::EXPEND_1);
-	Add_Expendables(CExpendables::EXPEND_2);
-	Add_Expendables(CExpendables::EXPEND_3);
-	Add_Expendables(CExpendables::EXPEND_4);
+
+	Add_MultiExpendables(CExpendables::EXPEND_1, 15);
+	Add_MultiExpendables(CExpendables::EXPEND_2, 15);
+	Add_MultiExpendables(CExpendables::EXPEND_3, 15);
+	Add_MultiExpendables(CExpendables::EXPEND_4, 20);
+
+
 	return NOERROR;
 }
 
@@ -69,15 +76,11 @@ _int CExpendables_Inven::Update_GameObject(_double TimeDelta)
 {
 	CUI::Update_GameObject(TimeDelta);
 
-	if (g_pInput_Device->Key_Up(DIK_1))
-		m_bIsActive = !m_bIsActive;
-
 	m_pRendererCom->Add_RenderList(RENDER_UI, this);
 
 	D3DXMatrixOrthoLH(&m_matProj, WINCX, WINCY, 0.f, 1.0f);
-
-	if (m_bIsActive)
-		Click_Inven();
+	
+	Click_Inven();
 
 	for (auto& pSlot : m_vecSlot)
 	{
@@ -185,6 +188,8 @@ HRESULT CExpendables_Inven::SetUp_ConstantTable()
 
 void CExpendables_Inven::Click_Inven()
 {
+	if (!m_bIsActive)
+		return;
 
 	for (auto& pExpendSlot : m_vecSlot)
 	{
@@ -192,6 +197,8 @@ void CExpendables_Inven::Click_Inven()
 		{
 			if (!pExpendSlot->Get_Select() && pExpendSlot->Get_Size() > 0 && g_pInput_Device->Get_DIMouseState(CInput_Device::DIM_LB))
 			{
+				if (m_vecQuickSlot.size() == 8)
+					return;
 				pExpendSlot->Set_Select(true);
 				
 				m_vecQuickSlot.push_back(pExpendSlot);
@@ -205,10 +212,9 @@ void CExpendables_Inven::Click_Inven()
 		}
 	}
 
-	if (g_pInput_Device->Key_Up(DIK_6))
-		Add_Expendables(CExpendables::EXPEND_1);
-	if (g_pInput_Device->Key_Up(DIK_7))
-		Add_Expendables(CExpendables::EXPEND_1);
+
+	
+
 }
 
 void CExpendables_Inven::Add_Expendables(CExpendables::EXPEND_TYPE eType)
@@ -225,6 +231,12 @@ void CExpendables_Inven::Add_Expendables(CExpendables::EXPEND_TYPE eType)
 	Load_Expendables(pExpendables, 0);
 }
 
+void CExpendables_Inven::Add_MultiExpendables(CExpendables::EXPEND_TYPE eType, _uint iCnt)
+{
+	for (_uint i = 0; i < iCnt; ++i)
+		Add_Expendables(eType);
+}
+
 void CExpendables_Inven::Delete_QuickSlot(CExpendables_Slot* pSlot)
 {
 	for (_uint i = 0; i < m_vecQuickSlot.size(); ++i)
@@ -239,16 +251,80 @@ void CExpendables_Inven::Delete_QuickSlot(CExpendables_Slot* pSlot)
 	return;
 }
 
+void CExpendables_Inven::Use_Expendableas(CExpendables_Slot * pSlot)
+{
+	if (nullptr == pSlot)
+		return;
+	if (pSlot->Get_Size() == 0 ||
+		m_vecQuickSlot.size() == 0)
+		return;
+
+	pSlot->Delete_Item();
+
+	for (_uint i = 0; i < m_vecQuickSlot.size(); ++i)
+	{
+
+		if (m_vecQuickSlot[i] == pSlot &&
+			m_vecQuickSlot[i]->Get_Size() == 0)
+		{
+			m_vecQuickSlot.erase(m_vecQuickSlot.begin() + i);
+			m_vecQuickSlot.shrink_to_fit();
+			break;
+		}
+	}
+
+	_ulong idx = 0;
+	for (auto& pExSlot : m_vecSlot)
+	{
+		if (pExSlot->Get_Size() == 0)
+		{
+			m_vecSlot.erase(m_vecSlot.begin() + idx);
+			m_vecSlot.shrink_to_fit();
+			break;
+		}
+		
+		++idx;
+	}
+
+	_ulong Idx = 0;
+	for (auto& pExSlot : m_vecSlot)
+	{
+		pExSlot->Set_UI_Pos(m_vecUI_DESC[Idx]->fPosX, m_vecUI_DESC[Idx]->fPosY);
+		++Idx;
+	}
+}
+
 void CExpendables_Inven::Sell_Expendables(_uint iDelete)
 {
+	_uint idx = 0;
 	for (auto& pSlot : m_vecSlot)
 	{
 		if (pSlot->Pt_InRect())
 		{
+			if (iDelete > pSlot->Get_Size())
+				return;
 			for (_uint i = 0; i < iDelete; ++i)
+			{
 				pSlot->Delete_Item();
-			return;
+			}
+					
+			if (pSlot->Get_Size() == 0)
+			{
+				m_vecSlot.erase(m_vecSlot.begin() + idx);
+				m_vecSlot.shrink_to_fit();
+				break;
+			}
 		}
+		
+		++idx;
+	}
+
+	_ulong Idx = 0;
+	for (auto& pSlot : m_vecSlot)
+	{
+
+		pSlot->Set_UI_Pos(m_vecUI_DESC[Idx]->fPosX, m_vecUI_DESC[Idx]->fPosY);
+		++Idx;
 	}
 }
